move browser related functions to a separate file
