+++ slackauth.go
@@ -5,14 +5,9 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
 	"time"
 
 	"github.com/go-rod/rod"
-	"github.com/go-rod/rod/lib/launcher"
 	"github.com/go-rod/rod/lib/proto"
 )
 
@@ -211,96 +206,4 @@ var sameSiteMap = map[proto.NetworkCookieSameSite]http.SameSite{
 	proto.NetworkCookieSameSiteNone:   http.SameSiteNoneMode,
 	proto.NetworkCookieSameSiteLax:    http.SameSiteLaxMode,
 	proto.NetworkCookieSameSiteStrict: http.SameSiteStrictMode,
-}
-
-func browserLauncher(headless bool) *launcher.Launcher {
-	var l *launcher.Launcher
-	if binpath, ok := lookPath(); ok {
-		l = launcher.New().
-			Bin(binpath).
-			Headless(headless).
-			Leakless(isLeaklessEnabled). // Causes false positive on Windows, see #260
-			Devtools(false)
-	} else {
-		l = launcher.New().
-			Leakless(isLeaklessEnabled). // Causes false positive on Windows, see #260
-			Headless(headless).
-			Devtools(false)
-	}
-	return l
-}
-
-// lookPath is extended launcher.LookPath that includes support for Brave
-// browser.
-//
-// (c) MIT license: Copyright 2019 Yad Smood
-func lookPath() (found string, has bool) {
-	list := map[string][]string{
-		"darwin": {
-			"/Applications/Brave Browser.app/Contents/MacOS/Brave Browser",
-			"/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
-			"/Applications/Chromium.app/Contents/MacOS/Chromium",
-			"/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge",
-			"/Applications/Google Chrome Canary.app/Contents/MacOS/Google Chrome Canary",
-			"/usr/bin/google-chrome-stable",
-			"/usr/bin/google-chrome",
-			"/usr/bin/chromium",
-			"/usr/bin/chromium-browser",
-		},
-		"linux": {
-			"brave-browser",
-			"chrome",
-			"google-chrome",
-			"/usr/bin/brave-browser",
-			"/usr/bin/google-chrome",
-			"microsoft-edge",
-			"/usr/bin/microsoft-edge",
-			"chromium",
-			"chromium-browser",
-			"/usr/bin/google-chrome-stable",
-			"/usr/bin/chromium",
-			"/usr/bin/chromium-browser",
-			"/snap/bin/chromium",
-			"/data/data/com.termux/files/usr/bin/chromium-browser",
-		},
-		"openbsd": {
-			"chrome",
-			"chromium",
-		},
-		"windows": append([]string{"chrome", "edge"}, expandWindowsExePaths(
-			`BraveSoftware\Brave-Browser\Application\brave.exe`,
-			`Google\Chrome\Application\chrome.exe`,
-			`Chromium\Application\chrome.exe`,
-			`Microsoft\Edge\Application\msedge.exe`,
-		)...),
-	}[runtime.GOOS]
-
-	for _, path := range list {
-		var err error
-		found, err = exec.LookPath(path)
-		has = err == nil
-		if has {
-			break
-		}
-	}
-
-	return
-}
-
-// expandWindowsExePaths is a verbatim copy of the function from rod's
-// browser.go.
-//
-// (c) MIT license: Copyright 2019 Yad Smood
-func expandWindowsExePaths(list ...string) []string {
-	newList := []string{}
-	for _, p := range list {
-		newList = append(
-			newList,
-			filepath.Join(os.Getenv("ProgramFiles"), p),
-			filepath.Join(os.Getenv("ProgramFiles(x86)"), p),
-			filepath.Join(os.Getenv("LocalAppData"), p),
-		)
-	}
-
-	return newList
 }
